fix & feat: sua loi tang hinh quai oc muon hon, toi uu toc do hieu ung mua/la roi va bao ve mob render loop

+++ src/models/player/EffectSkin.java
@@ -156,8 +156,8 @@ public class EffectSkin {
                 int param = this.player.nPoint.tlHutHpMpXQ;
                 if (param > 0) {
                     if (!this.player.isDie() && Util.canDoWithTime(lastTimeXenHutHpKi, 5000)) {
-                        int hpHut = 0;
-                        int mpHut = 0;
+                        long hpHut = 0;
+                        long mpHut = 0;
                         List<Player> players = new ArrayList<>();
                         List<Player> playersMap = this.player.zone.getNotBosses();
                         for (Player pl : playersMap) {
@@ -180,8 +180,8 @@ public class EffectSkin {
                             }
                         }
                         for (Player pl : players) {
-                            int subHp = (int) ((long) pl.nPoint.hpMax * param / 100);
-                            int subMp = (int) ((long) pl.nPoint.mpMax * param / 100);
+                            long subHp = (long) pl.nPoint.hpMax * param / 100;
+                            long subMp = (long) pl.nPoint.mpMax * param / 100;
                             if (subHp >= pl.nPoint.hp) {
                                 subHp = pl.nPoint.hp - 1;
                             }
@@ -218,7 +218,7 @@ public class EffectSkin {
                             Player pl = playersMap.get(i);
                             if (pl != null && pl.nPoint != null && !this.player.equals(pl) && !pl.isBoss && !pl.isDie()
                                     && Util.getDistance(this.player, pl) <= 200) {
-                                int subHp = (int) ((long) pl.nPoint.hpMax * param / 100);
+                                long subHp = (long) pl.nPoint.hpMax * param / 100;
                                 if (subHp >= pl.nPoint.hp) {
                                     subHp = pl.nPoint.hp - 1;
                                 }
@@ -382,7 +382,7 @@ public class EffectSkin {
             if (this.player.effectSkill != null && this.player.effectSkill.isBinh && this.player.effectSkill.playerUseMafuba != null) {
                 if (Util.canDoWithTime(lastTimeMaPhongBa, 500) && this.player.effectSkill.playerUseMafuba.playerSkill != null) {
                     double param = this.player.effectSkill.playerUseMafuba.playerSkill.getSkillbyId(Skill.MA_PHONG_BA).point * (this.player.effectSkill.typeBinh == 0 ? 1 : 2);
-                    int subHp = (int) ((long) this.player.effectSkill.playerUseMafuba.nPoint.hpMax * param / 100);
+                    long subHp = (long) (this.player.effectSkill.playerUseMafuba.nPoint.hpMax * param / 100);
                     if (subHp >= this.player.nPoint.hp) {
                         subHp = Math.abs(this.player.nPoint.hp - 100);
                     }
